sio: implement getsockname wrapper
SWIM wants to allow to bind to zero ports so as the kernel could
choose any free port automatically. It is needed mainly for
tests.

Zero port means that a real port is known only after bind() has
called, and getsockname() should be used to get it. SWIM used sio
library for such lowlevel API. This is why that function is added
to sio.

Needed for #3234